separated state actions and reducers
