+++ static/admin/src/constants/index.js
@@ -2,9 +2,30 @@
  * Created by karolisrusenas on 06/04/2016.
  */
 //
+// ------------------------------------
+// Auth Constants
+// ------------------------------------
 export const LOGIN_USER_REQUEST = 'LOGIN_USER_REQUEST'
 export const LOGIN_USER_FAILURE = 'LOGIN_USER_FAILURE'
 export const LOGIN_USER_SUCCESS = 'LOGIN_USER_SUCCESS'
 export const LOGOUT_USER = 'LOGOUT_USER'
 export const FETCH_PROTECTED_DATA_REQUEST = 'FETCH_PROTECTED_DATA_REQUEST'
 export const RECEIVE_PROTECTED_DATA = 'RECEIVE_PROTECTED_DATA'
+
+// ------------------------------------
+// State Constants
+// ------------------------------------
+export const SET_MODE = 'SET_MODE'
+export const GET_MODE = 'GET_MODE'
+
+export const REQUEST_STATE = 'REQUEST_STATE'
+export const RECEIVE_STATE = 'RECEIVE_STATE'
+
+export const REQUEST_RECORDS_COUNT = 'REQUEST_RECORDS_COUNT'
+export const RECEIVE_RECORDS_COUNT = 'RECEIVE_RECORDS_COUNT'
+
+export const REQUEST_STATS = 'REQUEST_STATS'
+export const RECEIVE_STATS = 'RECEIVE_STATS'
+
+export const SET_REFRESH_ID = 'SET_REFRESH_ID'
+export const CLEAR_REFRESH_ID = 'CLEAR_REFRESH_ID'
