+++ static/admin/src/redux/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 import { routerReducer as router } from 'react-router-redux'
-import info from './modules/state'
+import info from './modules/reducers/state'
 import navigation from './modules/navigation'
 import auth from './modules/reducers/auth'
